requirements.txt
aioboto3==12.0.0
aiobotocore==2.7.0
aiofiles==23.2.1
aiohttp==3.9.0
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.11.0
boto3==1.28.64
botocore==1.31.64
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
diskcache==5.6.3
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
h2client==0.2.3
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
jmespath==1.0.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.4
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
praw==7.7.1
prawcore==2.4.0
psutil==5.9.6
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pympanim==0.0.8
PyPika==0.48.9
python-dateutil==2.8.2
pytypeutils==0.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
requests==2.31.0
rqdb==1.2.1
s3transfer==0.7.0
scipy==1.11.4
si-prefix==1.2.2
six==1.16.0
sniffio==1.3.0
stripe==7.5.0
tinycss2==1.2.1
tqdm==4.66.1
twilio==8.10.2
typing_extensions==4.8.0
update-checker==0.18.0
urllib3==2.0.7
webencodings==0.5.1
websocket-client==1.6.4
wrapt==1.16.0
yarl==1.9.3
